+++ src/recording/screen-recorder-factory.vala
@@ -14,21 +14,30 @@ namespace Peek.Recording {
     public static ScreenRecorder create_default_screen_recorder () throws PeekError {
       string recorder;
 
-      if (check_for_executable ("ffmpeg")) {
+      if (GnomeShellDbusRecorder.is_available ()) {
+        recorder = "gnome-shell";
+      } else if (FfmpegScreenRecorder.is_available ()) {
         recorder = "ffmpeg";
-      } else if (check_for_executable ("avconv")) {
+      } else if (AvconvScreenRecorder.is_available ()) {
         recorder = "avconv";
       } else {
         throw new PeekError.NO_SUITABLE_SCREEN_RECORDER (
           "No suitable screen recorder found");
       }
 
-      debug ("Using screen recorder %s", recorder);
+      stdout.printf ("Using screen recorder %s\n", recorder);
       return create_screen_recorder (recorder);
     }
 
     public static ScreenRecorder create_screen_recorder (string name) throws PeekError {
       switch (name) {
+        case "gnome-shell":
+          try {
+            return new GnomeShellDbusRecorder ();
+          } catch (IOError e) {
+            throw new PeekError.SCREEN_RECORDER_ERROR (
+              e.message);
+          }
         case "ffmpeg":
           return new FfmpegScreenRecorder ();
         case "avconv":
@@ -37,29 +46,6 @@ namespace Peek.Recording {
           throw new PeekError.UNKNOWN_SCREEN_RECORDER (
             "Unknown screen recorder " + name);
       }
-    }
-
-    private static bool check_for_executable (string executable) {
-      string[] args = {
-        "which", executable
-      };
-
-      int status;
-      string output;
-      string errorout;
-
-      try {
-        Process.spawn_sync (null, args, null,
-          SpawnFlags.SEARCH_PATH,
-          null, out output, out errorout, out status);
-        debug ("Looking for executable %s (%i): %s%s",
-          executable, status, output, errorout);
-        return Utils.is_exit_status_success (status);
-      } catch (SpawnError e) {
-        debug ("Error: %s", e.message);
-      }
-
-      return false;
     }
   }